+++ lib/main.dart
@@ -1,11 +1,21 @@
+import 'package:arena/screens/stats_screen.dart';
 import 'package:flutter/material.dart';
-import 'screens/main_page.dart';
+import 'screens/home_screen.dart';
+
+import 'widgets/styled_app_bar.dart';
+import 'widgets/styled_bottom_navigation_bar.dart';
 
 void main() {
   runApp(
     MaterialApp(
       title: 'Arena',
-      home: MainPage(),
+      debugShowCheckedModeBanner: false,
+      home: Scaffold(
+        backgroundColor: Colors.grey.shade200,
+        appBar: StyledAppBar(),
+        body: HomeScreen(),
+        bottomNavigationBar: StyledBottomNavigationBar(),
+      ),
     ),
   );
 }
