+++ lib/screens/stats_screen.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class StatsScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
